Move Json (de)serializer into core
as we use them for the FileOpérationRepository
